+++ packages/rax-compat/src/shared.ts
@@ -1,33 +1,58 @@
-import type { ReactElement } from 'react';
+import * as React from 'react';
+import { unmountComponentAtNode } from 'react-dom';
 
+import type { ReactElement } from 'react';
 type ChildrenElement<T> = ChildrenElement<T>[] | T;
+
+// ref: https://github.com/alibaba/rax/blob/master/packages/rax/src/vdom/instance.js#L11
+const INSTANCE_KEY = '_r';
+const REACT_ELEMENT_TYPE = Symbol.for('react.element');
 
 // Mocked `Rax.shared`.
 const shared = {
-  get Element(): any {
-    warningCompat('shared.Element');
-    return null;
+  Element(type: any, key: any, ref: any, props: any, owner: any) {
+    // ref: https://github.com/facebook/react/blob/main/packages/react/src/ReactElement.js#L149-L161
+    return {
+      // This tag allows us to uniquely identify this as a React Element
+      $$typeof: REACT_ELEMENT_TYPE,
+
+      // Built-in properties that belong on the element
+      type: type,
+      key: key,
+      ref: ref,
+      props: props,
+
+      // Record the component responsible for creating this element.
+      _owner: owner,
+    };
   },
-  get Host(): any {
-    warningCompat('shared.Host');
-    return null;
+  Host: {
+    get owner() {
+      return (React as any).__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner.current;
+    },
   },
-  get Instance(): any {
-    warningCompat('shared.Instance');
-    return null;
+  Instance: {
+    get(node: any) {
+      return Object.assign({}, node[INSTANCE_KEY], {
+        _internal: {
+          unmountComponent: () => {
+            node.parentNode?.removeChild(node);
+            return unmountComponentAtNode(node);
+          },
+        },
+      });
+    },
+    remove(node: any) {
+      node[INSTANCE_KEY] = null;
+    },
+    set(node: any, instance: any) {
+      if (!node[INSTANCE_KEY]) {
+        node[INSTANCE_KEY] = instance;
+      }
+    },
   },
   flattenChildren,
 };
-
-function warningCompat(message: string) {
-  let stack: string;
-  try {
-    throw new Error(`You are not allowed to use ${message}.`);
-  } catch (error) {
-    stack = error.stack;
-  }
-  console.error(`[RaxCompat] ${stack}`);
-}
 
 function flattenChildren(children: ChildrenElement<ReactElement>) {
   if (children == null) {